feat: add indexedstack to nested route to login process
